Fixed spurious serial input character on startup and cleaned up code
When you start up the serial port the character 0xFE is read for some reason. Reading on the web, it looks like the receiver
get a false start bit and starts a read. So now I start up the serial port, turn on the interrupts and then spin for a couple of
ms. That gives the bad character to load into the receive buffer. Then I clear the buffer and away we go.

I added back in a general enableIRQ to Serial startup. I thought this was causing problems before but now it all seems to
be working fine and it means you can start the Serial and Timer objects in any order.

I also added the ability to show non-printable characters sent to Serial::puts. It shows them in standard octal notation. Serial::write
doesn't do this transformation so you can use it to send binary data as needed