Got code working reliably with XMODEM code in place
Had lots of issues with the code just stopping to work sometimes. I THINK I isolated it to having 2 enableIRQ() calls, one on the
Timer and the other on the UART. I added disableIRQ() so you never try to enable interrupts when they are already enabled.
Seems to work, but I'm not confident with the fix