Fixed a timing related bug in sdcard.c
When DEBUG_SD was on to dump a bunch of debug info everything worked. Turn that off and it started getting
timeout errors. There was a certain spot that needed a delay. Not sure what it needed to wait for. But I added
a delay there and all is well with debugging off now