Made utilities and sdfs into libraries
Combined bootutil and util into a single utility and made that into a static library. Also added periph.c to that library.
But I'll probably clean that up in favor of the C++ versions in kernel. I also moved SDFS and its friends into a
separate library. Now util and sdfs can be used by both the bootloader and kernel