Got rid of periph.cpp, cleaned up and fixed interrupt bug
The Serial code assumed interrupts are all set up at startup, which is not the case for the bootloader.
And we shouldn't be using interrupts in the bootloader anyway. So The startup assembly code for each
has an interruptsSupported call, which returns false for bootloader and true for kernel. If false, all of the
interrupt code is skipped and Serial::read() does a simple busy loop until a character is available and
then reads it directly.

All the old, hardcoded peripheral code is gone now. Everything goes through GPIO, Serial or Timer classes