WIP - Improvements to printf
The printf I was using from Haribote project was pretty simple. It didn't support many formats. I found a really nice one
from here: https://github.com/mpaland/printf. The problem is that it's too big for the bootloader. Since it's not the
bootloader but the kernel that needs a fully functional printf, I split them. I now have a Print class in baremetal which
expects a vsnprintf-like function to be supplied by the higher level code. For bootloader, this is adapted from the one
in Haribote. For the kernel it's adapted from mpaland.

Everything mostly works and it doesn't crash. But there are a couple of bugs in the Haribote adapted code.