Got my new Allocator working
dlmalloc turned out to be too hard to get working. So I wrote my own simpler allocator.
My intent is to use the techniques from dlmalloc to write a more easy to understand and
debug allocator.

I discovered that global constructors were not running. So I added the logic to do that. I also
discovered that if you don't have an explicit constructor for a class with a global instance, the
ctor will not run. Not sure if this is a bug or if I just need a compiler flag.

Since finding this out, I may go back and work on dlmalloc some more. But I'm really liking the
direction with my new allocator.