WIP - Getting FAT table reading going
Reading the FAT table to go to the next "cluster" works for directory reading, but not for file reading.
And it only works when the cluster size is 1 block. To this end, I've made strong types for Block and
Cluster so I can work out where each one is expected to be used.