WIP - Make RawFile into a virtual class
This will get rid of FileInfo as a way to pass info from the device specific
to device independent world. There will be a FAT32RawFile and a set
of virtual functions to retrieve information such as name and size.

The notion of a baseBlock will be hidden in the specialization of RawFile.
So we can make that concept device specific and, for instance, store a
base cluster for FAT32. It will also allow FAT32 member variables for
caching the last accessed FAT entry to make crossing cluster boundaries
more efficient.

I'll also work on the implementation of FAT traversal in general in this branch