Implemented FAT traversal
Moved FAT32RawFile and FAT32DirectoryIterator to their own files. FAT32DirectoryIterator now opens a directory file so
all file reads go through FAT32RawFile::read, which is where the FAT logic is.

Implemented all cases of file access. If the access is in the same cluster as the last access, we use the same physical cluster number.
Otherwise, we search through the cluster chain for the physical cluster that matches the desired logical cluster. Added an optimization
where, if the logical cluster is > the last accessed logical cluster the search starts from the last physical cluster.