WIP - Getting file creation working
Added logic to extend the size of the directory. Added file create, which
allocates a block from the FAT, finds a new empty directory entry and
creates an entry with the file name, a single block and a size of 0.

Also templatized strong types Cluster and Block.

Works on Mac, but hangs on RPi