The SDCard driver needs buffers aligned on 4 byte boundaries
All raw buffers now have an alignment directive. This leaks up into the File class because it has a
buffer passed directly to the RawFile interface. I've required that interface to have a 4 byte alignment
because even if it is not strictly true for some drivers, it will always be more efficient to align to 4 bytes
on a 32 bit architecture. Should it be 8 byte alignment to account for 64 bit architectures?