WIP - Implemented new version of printf
This is a hand written version, playing with some ideas after discussion with Jeff. This version is
used in both the bootloader and kernel. The next step is to reduce the size for the bootloader.
It is 1.8k bigger than before. I can remove float support and some other features to get the size
down. The good news is that it is 1.3k smaller for the kernel.

Next I will look at Jeffs code and see how we can collaborate on the BEST. PRINTF. EVARRR!!!