WIP - Moved xmodem code into a class and make it standalone
Now XYModem is a class. You instantiate it with lambda functions for serial I/O and system time
and it will do its work and call the passed lambda with each byte output. It no longer deals with
the address in the lambda. That is all handled by the caller