Code restructuring for Arduino IDE compatibility
Arduino has some really restrictive rules on including libraries. First of all, they have to be in the sketch/libraries path.
So I symlinked baremetal to libraries. Then it searches all libraries to resolve header files, so name collisions are
possible. For instance, there is a GPIO.h in the arduino header files. To fix this you can use a trick to put a dummy header file
in the root of your library directory. In this case I used bare.h. Then you make a subdirectory with this same root name (I used
'bare'), and put all the headers in there. Then for every source file using the bare library you do this:

    #include "bare.h"
    #include "bare/GPIO.h"

That works with macOS, the ARM makefiles and the Arduino IDE. In order to avoid having an actual empty bare.h
file, I renamed util.h to bare.h and put all the basic utility functions in there. I also made a corresponding change to
util.cpp. It is now bare.cpp.