Big restructuring to start make the whole system crossplatform
The goal is to get rid of all #ifdefs for platforms except in bare.h. Currently have
platform specific files for a few things but not all. Builds on macOS and RPi through
the Makefiles. ESP is structured correctly but there are some errors in the .ino file.

Still lots of work to get platform specific files separated, but this is a start