Fixed XYModem receive - both now work
There's a quirk in XModem receive in the bootloader. It takes a while for the receive to finish so you miss
the first few lines of output text when the new kernel boots up. Not sure if this is a Serial.app issue or what.
For now XModem works fine for the bootloader. YModem works correctly for the kernel 'put' command