Fix a bug in printf
Different implementations handle va_list differently. Some implementations represent it as an array. Passing it to a function passes
it as a pointer, so updates to it are returned. Others represent it as a struct, so it is passed by value and updates are not reflected
in the caller. But you can't pass it by reference because you can't take a reference to an array. So the cross-platform way I'm using is
to make a struct with a va_list in it, copying from the incoming va_list into the one in the struct and then passing that by reference.