Added a loop to the SPI test.
Now it performs the test 10 times and the master sends a status to the slave saying it's done.
Trying to see if data transfer is reliable. Next I'll compare the strings returned to make sure
they're correct and then bump up the data rate to see how fast I can get