Implemented many more API calls for WiFiSpi
This includes a new template method for sending params across the wire. I'm using
C++11 variadic templates to write a function that can take any number and type of
parameters. Of course to make this work, there needs to be a sendParam() function
that supported the type you want. So far I've implemented uint8_t and const char*.
More will be added as needed.