Changed stack setup a bit
I was actually not setting the SVC (SWI) stack properly. Now the FIQ stack is at 0x1000-0x800. I moved it there
since FIQ is probably going to be the smallest stack user. That puts the IRQ stack at 0x2000, the Abort and Undefined stack
at 0x3000 and the SVC stack at 0x4000. The System stack is still at 0x100000 like it was before