WIP - Added loading of executable to ELFLoader
A block of memory is created to hold the user program. Theoretically you could
jump to the _entryPoint offset in this block and it would execute the user program.

Still not relocating and the executable memory is allocated in kernel space. Plus of course
you can't execute ARM code on a Mac - yet.