Moved the user program from 0x8000 to 0x80
Linker defaults to 0x8000. I tried forcing it to 0 but that causes
a block of 64K to be put at the beginning. I think this has something
to do with mmap, but I'm not sure. I was able to put the start addr
at 0x80 and that resulted in a 700 byte executable.